Views and monitors secured

+++ src/main/java/cz/etnetera/reesmo/controller/monitor/MonitorController.java
@@ -1,6 +1,7 @@
 package cz.etnetera.reesmo.controller.monitor;
 
 import cz.etnetera.reesmo.controller.MenuActivityController;
+import cz.etnetera.reesmo.http.ControllerModel;
 import cz.etnetera.reesmo.message.Localizer;
 import cz.etnetera.reesmo.model.datatables.monitor.MonitorDT;
 import cz.etnetera.reesmo.model.form.monitor.MonitorCommand;
@@ -10,10 +11,12 @@ import cz.etnetera.reesmo.model.mongodb.monitoring.FlatlineMonitoring;
 import cz.etnetera.reesmo.model.mongodb.monitoring.FrequencyMonitoring;
 import cz.etnetera.reesmo.model.mongodb.monitoring.Monitoring;
 import cz.etnetera.reesmo.model.mongodb.project.Project;
+import cz.etnetera.reesmo.model.mongodb.user.Permission;
 import cz.etnetera.reesmo.model.mongodb.view.View;
 import cz.etnetera.reesmo.repository.mongodb.monitor.MonitorRepository;
 import cz.etnetera.reesmo.repository.mongodb.project.ProjectRepository;
 import cz.etnetera.reesmo.repository.mongodb.view.ViewRepository;
+import cz.etnetera.reesmo.user.UserManager;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -28,6 +31,8 @@ import java.util.concurrent.TimeUnit;
 @Controller
 public class MonitorController implements MenuActivityController {
 
+    @Autowired
+    private UserManager userManager;
 
     @Autowired
     MonitorRepository monitorRepository;
@@ -54,8 +59,12 @@ public class MonitorController implements MenuActivityController {
     @RequestMapping(value = "/monitor/{monitorId}", method = RequestMethod.GET)
     public String monitorDetail(@PathVariable String monitorId, Model model, Locale locale) {
         Monitoring monitor = monitorRepository.findOne(monitorId);
+        ControllerModel.exists(monitor, Monitoring.class);
+        Project project = projectRepository.findOne(monitor.getProjectId());
+        ControllerModel.exists(project, Project.class);
+        project.checkUserPermission(userManager.requireUser(), Permission.BASIC);
         View view = viewRepository.findOne(monitor.getViewId());
-        Project project = projectRepository.findOne(view.getProjectId());
+        ControllerModel.exists(view, View.class);
         MonitorDT monitorDT = new MonitorDT(monitor, localizer, locale);
         monitorDT.setViewName(view.getName());
         model.addAttribute("project", project);
@@ -67,7 +76,10 @@ public class MonitorController implements MenuActivityController {
     @RequestMapping(value = "/view/{viewId}/monitor/frequency/create", method = RequestMethod.GET)
     public String createFrequencyMonitorForm(@PathVariable String viewId, Model model) {
         View view = viewRepository.findOne(viewId);
+        ControllerModel.exists(view, View.class);
         Project project = projectRepository.findOne(view.getProjectId());
+        ControllerModel.exists(project, Project.class);
+        project.checkUserPermission(userManager.requireUser(), Permission.ADMIN);
         model.addAttribute("monitorCommand", new MonitorCommand(TimeUnit.HOURS, 0, 0, true));
         model.addAttribute("project", project);
         model.addAttribute("view", view);
@@ -76,9 +88,12 @@ public class MonitorController implements MenuActivityController {
 
     @RequestMapping(value = "/view/{viewId}/monitor/frequency/create", method = RequestMethod.POST)
     public String createFrequencyMonitor(Model model, @PathVariable String viewId, @Valid @ModelAttribute MonitorCommand monitorCommand, BindingResult result) {
+        View view = viewRepository.findOne(viewId);
+        ControllerModel.exists(view, View.class);
+        Project project = projectRepository.findOne(view.getProjectId());
+        ControllerModel.exists(project, Project.class);
+        project.checkUserPermission(userManager.requireUser(), Permission.ADMIN);
         if (result.hasErrors()){
-            View view = viewRepository.findOne(viewId);
-            Project project = projectRepository.findOne(view.getProjectId());
             model.addAttribute("project", project);
             model.addAttribute("view", view);
             return "page/monitor/frequencyMonitorCreate";
@@ -86,6 +101,7 @@ public class MonitorController implements MenuActivityController {
         FrequencyMonitoring frequencyMonitoring = new FrequencyMonitoring();
         monitorCommand.toMonitor(frequencyMonitoring);
         frequencyMonitoring.setViewId(viewId);
+        frequencyMonitoring.setProjectId(project.getId());
         monitorRepository.save(frequencyMonitoring);
         return "redirect:/view/" + viewId;
     }
@@ -93,7 +109,10 @@ public class MonitorController implements MenuActivityController {
     @RequestMapping(value = "/view/{viewId}/monitor/flatline/create", method = RequestMethod.GET)
     public String createFlatlineMonitorForm(@PathVariable String viewId, Model model) {
         View view = viewRepository.findOne(viewId);
+        ControllerModel.exists(view, View.class);
         Project project = projectRepository.findOne(view.getProjectId());
+        ControllerModel.exists(project, Project.class);
+        project.checkUserPermission(userManager.requireUser(), Permission.ADMIN);
         model.addAttribute("monitorCommand", new MonitorCommand(TimeUnit.HOURS, 0, 0, true));
         model.addAttribute("project", project);
         model.addAttribute("view", view);
@@ -102,9 +121,12 @@ public class MonitorController implements MenuActivityController {
 
     @RequestMapping(value = "/view/{viewId}/monitor/flatline/create", method = RequestMethod.POST)
     public String createFlatlineMonitor(Model model, @PathVariable String viewId, @Valid @ModelAttribute MonitorCommand monitorCommand, BindingResult result) {
+        View view = viewRepository.findOne(viewId);
+        ControllerModel.exists(view, View.class);
+        Project project = projectRepository.findOne(view.getProjectId());
+        ControllerModel.exists(project, Project.class);
+        project.checkUserPermission(userManager.requireUser(), Permission.ADMIN);
         if (result.hasErrors()){
-            View view = viewRepository.findOne(viewId);
-            Project project = projectRepository.findOne(view.getProjectId());
             model.addAttribute("project", project);
             model.addAttribute("view", view);
             return "page/monitor/frequencyMonitorCreate";
@@ -112,6 +134,7 @@ public class MonitorController implements MenuActivityController {
         FlatlineMonitoring flatlineMonitoring = new FlatlineMonitoring();
         monitorCommand.toMonitor(flatlineMonitoring);
         flatlineMonitoring.setViewId(viewId);
+        flatlineMonitoring.setProjectId(project.getId());
         monitorRepository.save(flatlineMonitoring);
         return "redirect:/view/" + viewId;
     }
@@ -119,7 +142,10 @@ public class MonitorController implements MenuActivityController {
     @RequestMapping(value = "/view/{viewId}/monitor/any/create", method = RequestMethod.GET)
     public String createAnyMonitorForm(@PathVariable String viewId, Model model) {
         View view = viewRepository.findOne(viewId);
+        ControllerModel.exists(view, View.class);
         Project project = projectRepository.findOne(view.getProjectId());
+        ControllerModel.exists(project, Project.class);
+        project.checkUserPermission(userManager.requireUser(), Permission.ADMIN);
         model.addAttribute("monitorCommand", new MonitorCommand(TimeUnit.HOURS, 0, 0, true));
         model.addAttribute("project", project);
         model.addAttribute("view", view);
@@ -127,9 +153,20 @@ public class MonitorController implements MenuActivityController {
     }
 
     @RequestMapping(value = "/view/{viewId}/monitor/any/create", method = RequestMethod.POST)
-    public String createAnyMonitor(Model model, @PathVariable String viewId, @ModelAttribute MonitorCommand monitorCommand) {
+    public String createAnyMonitor(Model model, @PathVariable String viewId, @ModelAttribute MonitorCommand monitorCommand, BindingResult result) {
+        View view = viewRepository.findOne(viewId);
+        ControllerModel.exists(view, View.class);
+        Project project = projectRepository.findOne(view.getProjectId());
+        ControllerModel.exists(project, Project.class);
+        project.checkUserPermission(userManager.requireUser(), Permission.ADMIN);
+        if (result.hasErrors()){
+            model.addAttribute("project", project);
+            model.addAttribute("view", view);
+            return "page/monitor/anyMonitorCreate";
+        }
         AnyMonitoring anyMonitoring = new AnyMonitoring();
         anyMonitoring.setViewId(viewId);
+        anyMonitoring.setProjectId(project.getId());
         anyMonitoring.setEnabled(monitorCommand.isEnabled());
         monitorRepository.save(anyMonitoring);
         return "redirect:/view/" + viewId;
@@ -138,8 +175,12 @@ public class MonitorController implements MenuActivityController {
     @RequestMapping(value = "/monitor/delete/{monitorId}", method = RequestMethod.GET)
     public String deleteMonitor(Model model, @PathVariable String monitorId) {
         Monitoring monitor = monitorRepository.findOne(monitorId);
+        ControllerModel.exists(monitor, Monitoring.class);
+        Project project = projectRepository.findOne(monitor.getProjectId());
+        ControllerModel.exists(project, Project.class);
+        project.checkUserPermission(userManager.requireUser(), Permission.ADMIN);
         View view = viewRepository.findOne(monitor.getViewId());
-        Project project = projectRepository.findOne(view.getProjectId());
+        ControllerModel.exists(view, View.class);
         model.addAttribute("monitor", monitor);
         model.addAttribute("monitorType", monitor.getClass().getSimpleName());
         model.addAttribute("project", project);
@@ -150,7 +191,12 @@ public class MonitorController implements MenuActivityController {
     @RequestMapping(value = "/monitor/delete/{monitorId}", method = RequestMethod.POST)
     public String deleteMonitor(@PathVariable String monitorId) {
         Monitoring monitor = monitorRepository.findOne(monitorId);
+        ControllerModel.exists(monitor, Monitoring.class);
+        Project project = projectRepository.findOne(monitor.getProjectId());
+        ControllerModel.exists(project, Project.class);
+        project.checkUserPermission(userManager.requireUser(), Permission.ADMIN);
         View view = viewRepository.findOne(monitor.getViewId());
+        ControllerModel.exists(view, View.class);
         monitorRepository.deleteMonitorAndNotifiers(monitorId);
         return "redirect:/view/" + view.getId();
     }
@@ -158,8 +204,12 @@ public class MonitorController implements MenuActivityController {
     @RequestMapping(value = "/monitor/edit/{monitorId}", method = RequestMethod.GET)
     public String editMonitor(Model model, @PathVariable String monitorId) {
         Monitoring monitor = monitorRepository.findOne(monitorId);
+        ControllerModel.exists(monitor, Monitoring.class);
+        Project project = projectRepository.findOne(monitor.getProjectId());
+        ControllerModel.exists(project, Project.class);
+        project.checkUserPermission(userManager.requireUser(), Permission.ADMIN);
         View view = viewRepository.findOne(monitor.getViewId());
-        Project project = projectRepository.findOne(view.getProjectId());
+        ControllerModel.exists(view, View.class);
         MonitorCommand monitorCommand = new MonitorCommand(monitor);
         model.addAttribute("monitorCommand", monitorCommand);
         model.addAttribute("project", project);
@@ -168,12 +218,22 @@ public class MonitorController implements MenuActivityController {
     }
 
     @RequestMapping(value = "/monitor/edit/{monitorId}", method = RequestMethod.POST)
-    public String editMonitor(@PathVariable String monitorId, @ModelAttribute MonitorCommand monitorCommand) {
+    public String editMonitor(Model model, @PathVariable String monitorId, @ModelAttribute MonitorCommand monitorCommand, BindingResult result) {
         Monitoring monitor = monitorRepository.findOne(monitorId);
+        ControllerModel.exists(monitor, Monitoring.class);
+        Project project = projectRepository.findOne(monitor.getProjectId());
+        ControllerModel.exists(project, Project.class);
+        project.checkUserPermission(userManager.requireUser(), Permission.ADMIN);
+        if (result.hasErrors()){
+            View view = viewRepository.findOne(monitor.getViewId());
+            ControllerModel.exists(view, View.class);
+            model.addAttribute("project", project);
+            model.addAttribute("view", view);
+            return "page/monitor/monitorEdit";
+        }
         monitorCommand.toMonitor(monitor);
         monitorRepository.save(monitor);
         return "redirect:/monitor/" + monitorId;
     }
-
 
 }
